Hide BlockBuilder inside ProgramBuilder. Add docstrings. bin/fmt.
