+++ universal-application-tool-0.0.1/test/support/ResourceCreator.java
@@ -2,7 +2,6 @@ package support;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.util.Locale;
 import models.Applicant;
@@ -10,7 +9,6 @@ import models.Program;
 import models.Question;
 import play.inject.Injector;
 import services.Path;
-import services.program.BlockDefinition;
 import services.program.ProgramDefinition;
 import services.question.QuestionDefinition;
 import services.question.QuestionService;
